packages-news-summarizer/requirements.txt
feedparser>=6.0.0
python-dotenv>=0.19.0
requests>=2.25.0